+++ core/model/src/main/java/uk/ac/ebi/interpro/scan/model/raw/Hmmer2RawMatch.java
@@ -17,9 +17,9 @@ public abstract class Hmmer2RawMatch extends HmmerRawMatch {
 
     protected Hmmer2RawMatch(String sequenceIdentifier, String model,
                              String signatureLibraryName, String signatureLibraryRelease,
-                             long locationStart, long locationEnd,
+                             int locationStart, int locationEnd,
                              double evalue, double score,
-                             long hmmStart, long hmmEnd, String hmmBounds,
+                             int hmmStart, int hmmEnd, String hmmBounds,
                              double locationEvalue, double locationScore, String generator) {
         super(sequenceIdentifier, model, signatureLibraryName, signatureLibraryRelease, locationStart, locationEnd, evalue, score, hmmStart, hmmEnd, hmmBounds, locationScore, generator);
         this.locationEvalue = Math.log10(locationEvalue);
